# backend-golang/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/api/v1/article": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "查询文章",
                "parameters": [
                    {
                        "description": "查询文章",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ArticleQueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/article/add": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "新增文章",
                "parameters": [
                    {
                        "description": "x新增文章",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ArticleAddOrUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/book": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book"
                ],
                "summary": "查询书籍",
                "parameters": [
                    {
                        "description": "查询书籍",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.BookQueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/book/add": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book"
                ],
                "summary": "新增书籍",
                "parameters": [
                    {
                        "description": "新增书籍",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ArticleAddOrUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/classification": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Classification"
                ],
                "summary": "查询类别",
                "parameters": [
                    {
                        "description": "查询类别",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ClassificationQueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/classification/add": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Classification"
                ],
                "summary": "新增类别",
                "parameters": [
                    {
                        "description": "新增类别",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ClassificationKey"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/common/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户登录接口",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "request.ArticleAddOrUpdateRequest": {
            "type": "object",
            "properties": {
                "classification": {
                    "type": "integer"
                },
                "first": {
                    "type": "string"
                },
                "gmt_create": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "platform": {
                    "type": "string"
                },
                "second": {
                    "type": "string"
                },
                "third": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "request.ArticleQueryRequest": {
            "type": "object",
            "properties": {
                "classification": {
                    "type": "integer"
                },
                "currentPage": {
                    "type": "integer"
                },
                "first": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "platform": {
                    "type": "string"
                },
                "second": {
                    "type": "string"
                },
                "sorter": {
                    "type": "string"
                },
                "third": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "request.BookQueryRequest": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "type": "integer"
                },
                "first": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pageSize": {
                    "type": "integer"
                },
                "second": {
                    "type": "string"
                },
                "sorter": {
                    "type": "string"
                },
                "third": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "request.ClassificationKey": {
            "type": "object",
            "properties": {
                "first": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "second": {
                    "type": "string"
                },
                "third": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "request.ClassificationQueryRequest": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "type": "integer"
                },
                "first": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "second": {
                    "type": "string"
                },
                "sorter": {
                    "type": "string"
                },
                "third": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "request.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}